feat(layout): add dynamic layout loading
