+++ src/App.vue
@@ -1,56 +1,31 @@
 <template>
   <div :class="['app-container', popupStore.isOpen && 'scale']">
-    <LifePointsCounter :player="player1" />
-    <Board :board="player1.board" />
-
-    <Button @click="handleAttackAll">Attack with all</Button>
-    <Button @click="handleNextTurn">Next Turn</Button>
-
-    <Library @summon="handleSummon" />
+    <suspense>
+      <component :is="appStore.layout.component" />
+    </suspense>
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent } from "vue"
 
-import { LibraryEntry } from "./@types/library"
 import { useAppStore } from "./stores/appStore"
-import { TokenCreature } from "./entities/TokenCreature"
 
 import { hydratePlayerStore } from "./utils/hydratePlayerStore"
-import Board from "./components/Board.vue"
-import Button from "./components/base/Button.vue"
-import Library from "./components/Library.vue"
-import LifePointsCounter from "./components/LifePointsCounter.vue"
 import { usePopupStore } from "./stores/popupStore"
 
 export default defineComponent({
   name: "App",
-  components: {
-    Board,
-    Button,
-    Library,
-    LifePointsCounter,
-  },
+  components: {},
   setup() {
     const appStore = useAppStore()
     const player1 = appStore.players[0]
-    const handleAttackAll = () => player1.board.attackWithAll()
-    const handleNextTurn = () => player1.nextTurn()
 
     hydratePlayerStore(player1)
 
-    const handleSummon = (entry: LibraryEntry) => {
-      const newToken = new TokenCreature({
-        ...entry.tokenOptions,
-      })
-
-      player1.board.addToken(newToken)
-    }
-
     const popupStore = usePopupStore()
 
-    return { player1, appStore, popupStore, handleAttackAll, handleNextTurn, handleSummon }
+    return { appStore, popupStore }
   },
 })
 </script>
